+++ src/test/kotlin/com/zepben/ewb/streaming/get/QueryNetworkStateServiceTest.kt
@@ -20,7 +20,7 @@ import com.zepben.protobuf.ns.GetCurrentStatesResponse
 import com.zepben.protobuf.ns.SetCurrentStatesResponse
 import com.zepben.testutils.exception.ExpectException.Companion.expect
 import io.grpc.Status
-import io.grpc.StatusRuntimeException
+import io.grpc.StatusException
 import io.grpc.stub.StreamObserver
 import io.grpc.testing.GrpcCleanupRule
 import io.mockk.*
@@ -28,6 +28,7 @@ import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.*
 import org.junit.Rule
 import org.junit.jupiter.api.Test
+import org.slf4j.LoggerFactory
 import java.time.LocalDateTime
 
 internal class QueryNetworkStateServiceTest {
@@ -60,8 +61,8 @@ internal class QueryNetworkStateServiceTest {
     private val onCurrentStatesStatusInterface = mockk<QueryNetworkStateService.CurrentStatesStatusHandler> { justRun { handle(any()) } }
     private val onProcessingErrorInterface = mockk<QueryNetworkStateService.ProcessingErrorHandler> { justRun { handle(any()) } }
 
-    private val service = QueryNetworkStateService(onGetCurrentStates, onCurrentStatesStatus, onProcessingError)
-    private val serviceJava = QueryNetworkStateService(onGetCurrentStatesInterface, onCurrentStatesStatusInterface, onProcessingErrorInterface)
+    private val service = QueryNetworkStateService(onGetCurrentStates, onCurrentStatesStatus, LoggerFactory.getLogger(""), onProcessingError)
+    private val serviceJava = QueryNetworkStateService(onGetCurrentStatesInterface, onCurrentStatesStatusInterface, LoggerFactory.getLogger(""), onProcessingErrorInterface)
 
     @Test
     internal fun getCurrentStates() {
@@ -88,8 +89,8 @@ internal class QueryNetworkStateServiceTest {
         val responseError = slot<Throwable>()
         verifySequence { responseObserver.onError(capture(responseError)) }
 
-        assertThat(responseError.captured, instanceOf(StatusRuntimeException::class.java))
-        (responseError.captured as StatusRuntimeException).status.let {
+        assertThat(responseError.captured, instanceOf(StatusException::class.java))
+        (responseError.captured as StatusException).status.let {
             assertThat(it.code, equalTo(Status.UNKNOWN.code))
             assertThat(it.cause, equalTo(error))
         }